requirements.txt
python-telegram-bot
openai
pydantic-settings
loguru
httpx
python-dotenv

